Move publication page discovery to its own test
