+++ packages/framework/tests/Feature/PageCollectionTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use function copy;
 use Hyde\Foundation\HydeKernel;
 use Hyde\Foundation\PageCollection;
 use Hyde\Hyde;
@@ -13,8 +12,6 @@ use Hyde\Pages\DocumentationPage;
 use Hyde\Pages\HtmlPage;
 use Hyde\Pages\MarkdownPage;
 use Hyde\Pages\MarkdownPost;
-use Hyde\Publications\Models\PublicationListPage;
-use Hyde\Publications\Models\PublicationPage;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\File;
@@ -182,45 +179,5 @@ class PageCollectionTest extends TestCase
         $this->assertEquals(new DocumentationPage('foo'), $collection->get('.source/docs/foo.md'));
 
         File::deleteDirectory(Hyde::path('.source'));
-    }
-
-    public function test_publication_pages_are_discovered()
-    {
-        mkdir(Hyde::path('publication'));
-        $this->createPublication();
-
-        $collection = PageCollection::boot(Hyde::getInstance())->getPages();
-        $this->assertCount(4, $collection); // Default pages + publication index + publication page
-        $this->assertInstanceOf(PublicationPage::class, $collection->get('publication/foo.md'));
-
-        File::deleteDirectory(Hyde::path('publication'));
-    }
-
-    public function test_listing_pages_for_publications_are_discovered()
-    {
-        mkdir(Hyde::path('publication'));
-        $this->createPublication();
-
-        $this->assertInstanceOf(
-            PublicationListPage::class,
-            PageCollection::boot(Hyde::getInstance())->getPage('publication/index')
-        );
-
-        File::deleteDirectory(Hyde::path('publication'));
-    }
-
-    protected function createPublication(): void
-    {
-        copy(Hyde::path('tests/fixtures/test-publication-schema.json'), Hyde::path('publication/schema.json'));
-        file_put_contents(Hyde::path('publication/foo.md'),
-            '---
-__canonical: canonical
-__createdAt: 2022-11-16 11:32:52
-foo: bar
----
-
-Hello World!
-'
-        );
     }
 }
